pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "pdfwerks"
dynamic = ["version"]
dependencies = [
  "pymupdf>=1.25.0",
  "pillow>=11.0.0",
  "rich>=10.0.0",
  "prompt_toolkit>=3.0.0",
  "pyperclip>=1.8.0"
]
requires-python = ">=3.8"
authors = [{ name = "Adithya Menon R", email = "adukottan@gmail.com" }]
description = "A lightweight Python toolkit with multiple tools for PDF manipulation"
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["pdfwerks", "pdf", "pdf-management", "toolkit", "tui", "cli", "python"]

[project.optional-dependencies]
docs = ["mkdocs>=1.5.0", "mkdocs-material>=9.5.0"]
test = ["pytest>=8.3.0"]

[project.urls]
Repository = "https://github.com/adithya-menon-r/PDFwerks.git"
Documentation = "https://adithya-menon-r.github.io/PDFwerks"

[project.scripts]
pdfwerks = "pdfwerks.main:main"

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"
